+++ tests/unit/api/userManagement.spec.ts
@@ -2,14 +2,17 @@ import Student from "@/api/api_models/user_management/Student";
 import UserManagement from "@/api/UserManagement";
 import { Account } from "@/entities/Account";
 import { Role } from "@/entities/Role";
+import { store } from "@/use/store/store";
+import { MutationTypes } from "@/use/store/mutation-types";
+import { getRandomizedUserAndAuthUser } from "../../helper/Users";
 import { readFileSync } from "fs";
 import MachineUserAuthenticationManagement from "../../helper/MachineUserAuthenticationManagement";
-import { getRandomizedUserAndAuthUser } from "../../helper/Users";
 
 var userManagement: UserManagement;
-const pair = getRandomizedUserAndAuthUser(Role.STUDENT) as { student: Student; authUser: Account };
+const pair = getRandomizedUserAndAuthUser(Role.STUDENT) as { governmentId: string; student: Student; authUser: Account };
 const student = pair.student;
 const authUser = pair.authUser;
+const governmentId = pair.governmentId;
 
 const adminAuth = JSON.parse(readFileSync("tests/fixtures/logins/admin.json", "utf-8")) as { username: string; password: string };
 const studentAuth = JSON.parse(readFileSync("tests/fixtures/logins/student.json", "utf-8")) as { username: string; password: string };
@@ -30,7 +33,7 @@ beforeAll(async () => {
 });
 
 test("Create user", async () => {
-    const success = await userManagement.createUser(authUser, student);
+    const success = await userManagement.createUser(governmentId, authUser, student);
     expect(success.returnValue).toBe(true);
     await new Promise((r) => setTimeout(r, 1000));
 });
@@ -39,6 +42,7 @@ test("Get specific user", async () => {
     var result = false;
     const user = await userManagement.getSpecificUser(student.username);
     result = user.returnValue.username == student.username;
+    student.enrollmentIdSecret = user.returnValue.enrollmentIdSecret;
     expect(result).toBe(true);
 });